// src/modules/miniapi/types/nef.enums.ts
export enum NEFOperationType {
  NEF_AUTHENTICATION = 'Def19Sec9',
  AUTHENTICATION_WITH_5GS = 'Def115G1',
  CREATE_UE = '2',
  GET_UES = '3',
  NEF_LOCATION_SUBSCRIPTION = 'Def115G2',
  UE_PATH_LOSS = 'Def115G5',
  ACQUISITION_OF_RSRP = 'Def115G4',
  SERVING_CELL_INFO = 'Def115G6',
  HANDOVER = 'Def115G3',
  SUBSCRIBE_QOS_EVENT = 'Def115G7',
  E2E_SINGLE_UE_LATENCY_AND_THROUGHPUT = 'Def14Perf1',
  E2E_MULTIPLE_UE_LATENCY_AND_THROUGHPUT = 'Def14Perf2',
  MAX_HOPS = 'Def14Perf13',
  MAX_CONNECTIONS = 'Def14Perf11',
  NEF_CALLBACK_MAX_CONNECTIONS = 'Def14Perf7',
}